move CgmesControlArea to extensions package
