+++ cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/extensions/CgmesControlArea.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.cgmes.conversion.elements.areainterchange;
+package com.powsybl.cgmes.conversion.extensions;
 
 import java.util.HashSet;
 import java.util.Objects;
